+++ src/routers/Tutoria_routes.js
@@ -8,7 +8,11 @@ import {
   eliminarTutoria,
   listarInscripciones,
   inscribirTutoria,
-  desinscribirTutoria
+  desinscribirTutoria,
+  aceptarInscripcion,
+  rechazarInscripcion,
+  listarInscripcionesDocente,
+  listarInscripcionesEstudiante
 } from '../controllers/tutoria_controllers.js'
 
 const routerTutoria = Router()
@@ -21,5 +25,11 @@ routerTutoria.delete('/admin/tutoria/:id', verificarTokenJWT, eliminarTutoria)
 routerTutoria.get('/admin/tutoria/:id/inscripciones', verificarTokenJWT, listarInscripciones)
 routerTutoria.post('/admin/tutoria/:id/inscribir', verificarTokenJWT, inscribirTutoria)
 routerTutoria.delete('/admin/tutoria/:id/inscribir', verificarTokenJWT, desinscribirTutoria)
+
+routerTutoria.get('/docente/tutorias/inscripciones', verificarTokenJWT, listarInscripcionesDocente)
+routerTutoria.put('/docente/tutoria/:id/inscripcion/:inscripcionId/aceptar', verificarTokenJWT, aceptarInscripcion)
+routerTutoria.put('/docente/tutoria/:id/inscripcion/:inscripcionId/rechazar', verificarTokenJWT, rechazarInscripcion)
+
+routerTutoria.get('/estudiante/tutorias/inscripciones', verificarTokenJWT, listarInscripcionesEstudiante)
 
 export default routerTutoria